requirements.txt
python-dotenv
discord
Flask